+++ lib/src/serialazy.ts
@@ -33,7 +33,7 @@ export function deflate(serializable: any): JsonType {
  * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
  * @returns Serializable instance
  */
-export function inflate<T>(ctor: Constructor.Default<T>, serialized: JsonType): T {
+export function inflate<T>(ctor: Constructor<T>, serialized: JsonType): T {
 
     if (typeof(ctor) !== 'function') {
         throw new Error('Expecting a constructor function');